% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aovt.R
\name{aov.t}
\alias{aov.t}
\title{aov.t}
\usage{
aov.t(formula, ranfac = NULL)
}
\arguments{
\item{formula, ranfac}{}
}
\value{
aov(df,sumsq,meansq)
}
\description{
analysis of variance
}
\examples{
a<-as.factor(c(rep(1,3),rep(2,3),rep(3,3),rep(4,3)))
b<-as.factor(c(rep(c(1,2,3),4)))
y<-c(97.6,97.3,96.7,98.6,98.2,96.9,99,98,97.9,98,97.7,96.5)
z<-data.frame(a,b,y)

}
