% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.split.R
\name{design.split}
\alias{design.split}
\title{design.split}
\usage{
design.split(trt1, trt2, r = NULL, seed = 0, randomization = TRUE,
  std = FALSE)
}
\arguments{
\item{trt1, trt2, r, seed, randomization, std}{}
}
\value{
std.order, run.order, block, trt1, trt2
}
\description{
design of split's random number table
}
\examples{
t1<-c(1,2,3)
t2<-c(1,2,3)

}
