% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facdesign.r
\name{fac.design}
\alias{fac.design}
\title{fac.design}
\usage{
fac.design(levels, nVars = 0, r = 1, center = TRUE,
  factors = "none", varNames = NULL, seed = 0, std = FALSE)
}
\arguments{
\item{levels, nVars, r, center, factors, varNames, seed, std}{}
}
\value{
std.order, run.order, level of factor
}
\description{
randomized block design of Factorial design
}
\examples{
levels<-c(2,2,3)
r=2
std=TRUE

}
