% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honova3.R
\name{honova3}
\alias{honova3}
\title{honova3}
\usage{
honova3(formula, data, alpha, k)
}
\arguments{
\item{formula, data, alpha, k}{}
}
\value{
ci plot
}
\description{
two way anova (CI plot / no repetition / mixed model)
}
\examples{
a<-as.factor(c(rep(1,3),rep(2,3),rep(3,3)))
b<-as.factor(rep(c(1,2,3),3))
y<-c(13.1,12.9,13.4,12.4,12.7,12.5,12.3,12,12.2)
z<-data.frame(a,b,y)
j<-suppressWarnings(anova(lm(y~a+b)))

}
