% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactioneffect.R
\name{interaction.effect}
\alias{interaction.effect}
\title{interaction.effect}
\usage{
interaction.effect(data, fac1, fac2, y)
}
\arguments{
\item{data, fac1, fac2, y}{}
}
\value{
interaction plot
}
\description{
interaction plot
}
\examples{
A<-c(1,1,2,2,3,3)
A<-as.factor(c(rep(A,6)))
B<-as.factor(c(rep(1,6),rep(2,6),rep(3,6),rep(4,6),rep(5,6),rep(6,6)))
y<-c(305,302,322,325,320,322,335,337,350,348,342,344,366,364,326,324,338,336,372,374,330,330,348,348,376,373,327,330,350,350,348,350,310,308,330,328)
z<-data.frame(A,B,y)

}
