% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbbd.R
\name{kbbd}
\alias{kbbd}
\title{kbbd}
\usage{
kbbd(k, n0 = 4, block = (k == 4 | k == 5), randomize = TRUE, coding)
}
\arguments{
\item{k, n0, block, randomize, coding}{}
}
\value{
std.order, run.order, level of factor
}
\description{
design of box-behnken
}
\examples{
k<-3

}
