% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kccd.R
\name{kccd}
\alias{kccd}
\title{kccd}
\usage{
kccd(basis, generators, blocks = "Block", n0 = 4,
  alpha = "orthogonal", wbreps = 1, bbreps = 1, randomize = TRUE,
  inscribed = FALSE, coding, oneblock = FALSE)
}
\arguments{
\item{basis, generators, block, n0, alpha, wbreps, bbreps, randomize, inscribed, coding, oneblock}{}
}
\value{
std.order, run.order, level of factor
}
\description{
central composite design
}
\examples{
basis=3
randomize=FALSE
oneblock=TRUE

}
