% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdesign.R
\name{kdesign}
\alias{kdesign}
\title{kdesign}
\usage{
kdesign(design, randomize = TRUE, replicates = 1)
}
\arguments{
\item{design, randomize, replicates}{}
}
\value{
table, anova
}
\description{
random number table of orthogonal arrays
}
\examples{
design = "L9_3"
randomize = TRUE
replicates = 1

}
