% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksanova.R
\name{ksanova}
\alias{ksanova}
\title{ksanova}
\usage{
ksanova(formula, data, alpha)
}
\arguments{
\item{formula, data, alpha}{}
}
\value{
ci, plot
}
\description{
two way anova (no repetition / main effect and some interaction)
}
\examples{
a<-as.factor(c(rep(1,9),rep(2,9),rep(3,9)))
b<-as.factor(c(1,1,1,2,2,2,3,3,3,1,1,1,2,2,2,3,3,3,1,1,1,2,2,2,3,3,3))
c<-as.factor(rep(c(1,2,3),9))
y<-c(74,86,76,72,91,87,48,65,56,61,78,71,62,81,77,55,72,63,50,70,60,49,68,64,52,69,60)
z<-data.frame(a,b,c,y)

}
