% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnova1.R
\name{nnova1}
\alias{nnova1}
\title{nnova1}
\usage{
nnova1(formula, data, alpha, na.rm = TRUE, verdose = TRUE)
}
\arguments{
\item{formula, data, alpha, na.rm, verdose}{}
}
\value{
plot, anova, CI
}
\description{
one way anova (CI plot / no repetition)
}
\examples{
a1<-c(49,73,58,38,42)
a2<-c(31,40,43,44,34,20)
a3<-c(46,41,58,31,65)
a4<-c(45,73,76)
a <- c(a1,a2,a3,a4)
group<-c(1,1,1,1,1,2,2,2,2,2,2,3,3,3,3,3,4,4,4)
group<-as.factor(group)
z<-data.frame(a,group)

}
