% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optcom.R
\name{optcom}
\alias{optcom}
\title{optcom}
\usage{
optcom(data, fac, ranfac = NULL, x, alpha = 0.05)
}
\arguments{
\item{data, fac, ranfac, x, alpha}{}
}
\value{
optimum level combination, confidence level
}
\description{
Optimum level combination
}
\examples{
fac<-c("a", "b","c","a*c")
a<-as.factor(c(rep(1,9),rep(2,9),rep(3,9)))
a<-as.factor(c(rep(1,9),rep(2,9),rep(3,9)))
b<-as.factor(c(1,1,1,2,2,2,3,3,3,1,1,1,2,2,2,3,3,3,1,1,1,2,2,2,3,3,3))
c<-as.factor(rep(c(1,2,3),9))
y<-c(74,86,76,72,91,87,48,65,56,61,78,71,62,81,77,55,72,63,50,70,60,49,68,64,52,69,60)
z<-data.frame(a,b,c,y)

}
