% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual.aov.r
\name{qual.aov}
\alias{qual.aov}
\title{qual.aov}
\usage{
qual.aov(formula, data, r, i, j, fac)
}
\arguments{
\item{formula, data, r, i, j, fac}{}
}
\value{
table, anova
}
\description{
total qualitative data
}
\examples{
A<-c(1,1,2,2,3,3,4,4)
B<-rep(c("plus","minus"),4)
Y<-c(190.00,10.00,178.00,22.00,194.00,6.00,170.00,30.00)
k<-data.frame(A,B,Y)

}
