% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomlatin.R
\name{random.latin}
\alias{random.latin}
\title{random.latin}
\usage{
random.latin(n, greco = FALSE, seed = 0, std = FALSE)
}
\arguments{
\item{n, greco, seed, std}{}
}
\value{
std.order, run.order, level of factor
}
\description{
random number table(latin,graeco)
}
\examples{
n <- 3

}
