% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randommulti.R
\name{random.multi}
\alias{random.multi}
\title{random.multi}
\usage{
random.multi(fac, r = 1, seed = 0, std = FALSE)
}
\arguments{
\item{fac, r, seed, std}{}
}
\value{
std.order, run.order, r, factorlevel
}
\description{
multiple way random number table
}
\examples{
fac<-c(2,3)

}
