% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomone.R
\name{random.one}
\alias{random.one}
\title{random.one}
\usage{
random.one(fac, r = 1, seed = 0, std = FALSE)
}
\arguments{
\item{fac, r, seed, std}{}
}
\value{
std.order, run.order, r, level of factor
}
\description{
random number table(one way)
}
\examples{
A<-c("a","b","c")

}
